move command building api: utils.algebraic -> api.algebraic
