+++ src/main/java/carpet/api/algebraic/AbstractAdtMatcher.java
@@ -1,4 +1,4 @@
-package carpet.utils.algebraic;
+package carpet.api.algebraic;
 
 import java.util.ArrayList;
 import java.util.Collection;
