+++ src/main/java/carpet/api/algebraic/StringAdtMatcher.java
@@ -1,4 +1,4 @@
-package carpet.utils.algebraic;
+package carpet.api.algebraic;
 
 public class StringAdtMatcher extends SingleAdtMatcher<String> {
     @Override
